requirements/dev-requirements.txt
Sphinx-Substitution-Extensions==2020.9.30.0
VWS-Python-Mock==2021.3.27.1
VWS-Test-Fixtures==2021.11.5.1
autoflake==1.4
black==21.12b0
check-manifest==0.47
doc8==0.10.1
docker==5.0.3
flake8-commas==2.1.0
flake8-quotes==3.3.1
flake8==4.0.1
freezegun==1.1.0
furo==2021.11.23
homebrew-pypi-poet==0.10.0
isort==5.10.1
mypy==0.910
pip_check_reqs==2.3.2
pydocstyle==6.1.1
pygithub==1.55
pylint==2.12.2
pyroma==3.2
pytest-cov==3.0.0  # Measure code coverage
pytest==6.2.5
sphinx-click==3.0.2
sphinx==4.3.1
sphinx-inline-tabs==2021.8.17b10
sphinxcontrib.spelling==7.3.0
types-freezegun==1.1.3
types-setuptools==57.4.4
types-PyYAML==6.0.1
vulture==2.3
